javajs.async package
Async
AsyncFileChooser
AsyncDialog
AsyncColorChooser

better modal blocking for Chrome and AsyncFileChooser
